Found a solution to move fetch request and loading logic to loading container
