+++ src/client/actions/index.js
@@ -1,13 +1,5 @@
-export const START_LOADING = 'start_loading';
 export const FETCH_LEARNERS = 'fetch_learners';
 export const DONE_LOADING = 'done_loading';
-
-export function startLoading() {
-  return {
-    type: START_LOADING,
-    loading: true,
-  };
-}
 
 export function fetchLearners(allLearners) {
   return {
