+++ src/client/components/app/index.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Route, Switch } from 'react-router-dom';
-import TalentNavbar from '../../containers/talentNavbar';
+import TalentNavbar from '../talentNavbar';
 import NotFound from '../notFound';
 import SkillsSearch from '../../containers/skillsSearch';
 import ScrollToTop from '../scrollToTop';
