+++ src/client/components/talentNavbar/index.jsx
@@ -1,25 +1,8 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
 import { Link } from 'react-router-dom';
 import { Navbar } from 'react-bootstrap';
-import { startLoading, fetchLearners, doneLoading } from '../../actions';
-import axios from 'axios';
 
-class TalentNavbar extends Component {
-
-  componentWillMount() {
-    this.props.startLoading();
-    axios.get('http://localhost:3000/api/learners')
-    .then(response => response.data)
-    .then(data => this.props.fetchLearners(data))
-    .then(() => this.props.doneLoading())
-    .catch(error => {
-      this.props.doneLoading();
-      console.log('Error fetching and parsing data: ', error);
-      throw error;
-    });
-  }
+export default class TalentNavbar extends Component {
 
   render() {
     return (
@@ -57,9 +40,3 @@ class TalentNavbar extends Component {
     );
   }
 }
-
-function mapDispatchToProps(dispatch) {
-  return bindActionCreators({ startLoading, fetchLearners, doneLoading, }, dispatch);
-}
-
-export default connect(null, mapDispatchToProps)(TalentNavbar);
