+++ src/client/containers/loading/index.jsx
@@ -1,8 +1,23 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { doneLoading } from '../../actions';
+import { bindActionCreators } from 'redux';
+import { fetchLearners, doneLoading } from '../../actions';
+import axios from 'axios';
 
 class Loading extends Component {
+
+  componentDidMount() {
+    axios.get('http://localhost:3000/api/learners')
+    .then(response => response.data)
+    .then(data => this.props.fetchLearners(data))
+    .then(() => this.props.doneLoading())
+    .catch(error => {
+      this.props.doneLoading();
+      console.log('Error fetching and parsing data: ', error);
+      throw error;
+    });
+  }
+
   render() {
     return (
       <div>
@@ -18,4 +33,8 @@ function mapStateToProps({ guild }) {
   return { guild };
 }
 
-export default connect(mapStateToProps)(Loading);
+function mapDispatchToProps(dispatch) {
+  return bindActionCreators({ fetchLearners, doneLoading, }, dispatch);
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(Loading);
